[/]: Header: delete unnecessary code with scroll animation and replace them by css
