+++ app/globals.scss
@@ -15,6 +15,7 @@ html,
 body {
   max-width: 100vw;
   overflow-x: hidden;
+  scroll-behavior: smooth;
   background-color: #f9f9f9;
 }
