+++ app/not-found.tsx
@@ -1,10 +1,11 @@
-import Header from "@/components/UI/Header/Header.component";
+
 import NotFound from "@/components/pages/Blog/NotFound/NotFound.page";
+import Header from "@/components/pages/Header/Header.component";
 
 function notFound() {
   return (
     <>
-      <Header type="out" />
+      <Header/>
       <NotFound link={"/"} />
     </>
   );
